# ENH: optimize.linprog: vectorize `marg_bnds` loop