@nickodell 2025-03-16
Pushed a new change to fix a duplicate initialization. I haven't rebenchmarked this in detail; it could plausibly change the benchmark results by an amount on the order of 0.0014%.  (Here's the math: `_highs_wrapper()` spends about 0.7 percent of its time in `_get_marg_bnds()`. `_get_marg_bnds` spends about 0.2 of its time for each call to `np.zeros()`. So eliminating one call will save about (0.7 / 100 * 0.2 / 100) * 100 = 0.0014%.)