@nathanWolo 2025-04-08
I was investigating something that we are in running in production where a MIP that can be solved in just a few seconds when directly using highspy was taking >20 minutes using scipy.optimize.linprog. It seems like effectively all of that 20 minutes is being spent in this `marg_bnds` loop. I believe that there is something that is causing access to `solution.col_dual` and `basis.col_status` to be incredibly slow.

If I replace direct calls to the attributes in the loop with pre-allocated lists:

```python
    sol_col_dual_list = list(solution.col_dual)
    basis_col_status_list = list(basis.col_status)
    for ii in range(numcol):
        if basis_col_status_list[ii] == _h.HighsBasisStatus.kLower:
            marg_bnds[0, ii] = sol_col_dual_list[ii]
        elif basis_col_status_list[ii] == _h.HighsBasisStatus.kUpper:
            marg_bnds[1, ii] = sol_col_dual_list[ii]
```
The runtime is greatly decreased (from >20 minutes to under a second). For reference this is a MIP with approx 55,000 decision variables. This is on the latest release `1.15.2`.