@nickodell 2025-04-08
> I was investigating something that we are in running in production where a MIP that can be solved in just a few seconds when directly using highspy was taking >20 minutes using scipy.optimize.linprog. It seems like effectively all of that 20 minutes is being spent in this `marg_bnds` loop. I believe that there is something that is causing access to `solution.col_dual` and `basis.col_status` to be incredibly slow.
>
> [...]
>
>The runtime is greatly decreased (from >20 minutes to under a second). For reference this is a MIP with approx 55,000 decision variables. This is on the latest release `1.15.2`.

A fix for this issue has been merged already in #22660. It is planned for release 1.15.3. If you need a fix earlier than that, we have [nightly builds](https://anaconda.org/scientific-python-nightly-wheels/scipy) which contain the fix from that PR. However, nightly builds have a higher risk of bugs.

This PR is about a much smaller performance issue: the prior PR replaces an O(N^2) loop with one that is O(N). This PR merely makes that O(N) loop faster by doing more of it in NumPy. See prior PR for details.